requirements.txt
openai
sqlalchemy
python-dotenv
langchain
clean-text
pgvector
psycopg2-binary